@@ -2,37 +2,46 @@ name: mono_repo
 description: >-
   CLI tools to make it easier to manage a single source repository containing
   multiple Dart packages.
-version: 3.4.7
+version: 4.0.0-dev
 repository: https://github.com/google/mono_repo.dart
 
 environment:
-  sdk: '>=2.7.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
-  args: ^1.4.0
-  checked_yaml: ^1.0.0
+  args: ^2.0.0
+  checked_yaml: ^2.0.0
   collection: ^1.14.3
-  graphs: ^0.2.0
-  io: ^0.3.2+1
-  json_annotation: ^3.0.0
+  graphs: ^1.0.0
+  io: ^1.0.0
+  json_annotation: ^4.0.0
   meta: ^1.0.0
   path: ^1.4.1
-  pub_semver: ^1.3.2
-  pubspec_parse: ^0.1.0
-  yaml: ^2.2.0
+  pub_semver: ^2.0.0
+  pubspec_parse: ^1.0.0
+  yaml: ^3.0.0
 
 dev_dependencies:
   build_runner: ^1.0.0
-  build_test: '>=0.10.0 <2.0.0'
-  build_verify: ^1.0.0
+  build_test: ^2.0.0
+  build_verify: ^2.0.0
   build_version: ^2.0.0
-  json_serializable: ^3.2.0
+  json_serializable: ^4.0.0
   # Because we use the ignore_for_file build.yaml option
-  source_gen: ^0.9.5
+  source_gen: ^1.0.0
   term_glyph: ^1.0.0
-  test: ^1.3.0
-  test_descriptor: ^1.0.0
-  test_process: ^1.0.1
+  test: ^1.16.5
+  test_descriptor: ^2.0.0
+  test_process: ^2.0.0
 
 executables:
   mono_repo:
+
+dependency_overrides:
+  args: ^2.0.0
+  checked_yaml: ^2.0.0
+  graphs: ^1.0.0
+  io: ^1.0.0
+  pub_semver: ^2.0.0
+  pubspec_parse: ^1.0.0
+  yaml: ^3.0.0
