@@ -16,14 +16,16 @@ ConditionalStage _$ConditionalStageFromJson(Map json) {
         disallowNullValues: const ['name', 'if']);
     final val = ConditionalStage(
       $checkedConvert(json, 'name', (v) => v as String),
-      $checkedConvert(json, 'if', (v) => v as String),
+      $checkedConvert(json, 'if', (v) => v as String?),
     );
     return val;
   }, fieldKeyMap: const {'ifCondition': 'if'});
 }
 
 Map<String, dynamic> _$ConditionalStageToJson(ConditionalStage instance) {
-  final val = <String, dynamic>{};
+  final val = <String, dynamic>{
+    'name': instance.name,
+  };
 
   void writeNotNull(String key, dynamic value) {
     if (value != null) {
@@ -31,7 +33,6 @@ Map<String, dynamic> _$ConditionalStageToJson(ConditionalStage instance) {
     }
   }
 
-  writeNotNull('name', instance.name);
   writeNotNull('if', instance.ifCondition);
   return val;
 }
