@@ -18,14 +18,10 @@ CIJob _$CIJobFromJson(Map json) {
       $checkedConvert(
           json,
           'tasks',
-          (v) => (v as List)
-              ?.map((e) => e == null
-                  ? null
-                  : Task.fromJson((e as Map)?.map(
-                      (k, e) => MapEntry(k as String, e),
-                    )))
-              ?.toList()),
-      description: $checkedConvert(json, 'description', (v) => v as String),
+          (v) => (v as List<dynamic>)
+              .map((e) => Task.fromJson(Map<String, dynamic>.from(e as Map)))
+              .toList()),
+      description: $checkedConvert(json, 'description', (v) => v as String?),
     );
     return val;
   });
@@ -45,7 +41,7 @@ Map<String, dynamic> _$CIJobToJson(CIJob instance) {
   val['package'] = instance.package;
   val['sdk'] = instance.sdk;
   val['stageName'] = instance.stageName;
-  val['tasks'] = instance.tasks?.map((e) => e?.toJson())?.toList();
+  val['tasks'] = instance.tasks.map((e) => e.toJson()).toList();
   return val;
 }
 
@@ -53,14 +49,16 @@ Task _$TaskFromJson(Map json) {
   return $checkedNew('Task', json, () {
     final val = Task(
       $checkedConvert(json, 'name', (v) => v as String),
-      args: $checkedConvert(json, 'args', (v) => v as String),
+      args: $checkedConvert(json, 'args', (v) => v as String?),
     );
     return val;
   });
 }
 
 Map<String, dynamic> _$TaskToJson(Task instance) {
-  final val = <String, dynamic>{};
+  final val = <String, dynamic>{
+    'name': instance.name,
+  };
 
   void writeNotNull(String key, dynamic value) {
     if (value != null) {
@@ -68,7 +66,6 @@ Map<String, dynamic> _$TaskToJson(Task instance) {
     }
   }
 
-  writeNotNull('name', instance.name);
   writeNotNull('args', instance.args);
   return val;
 }
