@@ -13,16 +13,16 @@ RawConfig _$RawConfigFromJson(Map json) {
     $checkKeys(json, allowedKeys: const ['os', 'dart', 'stages', 'cache']);
     final val = RawConfig(
       $checkedConvert(json, 'os',
-              (v) => (v as List)?.map((e) => e as String)?.toList()) ??
+              (v) => (v as List<dynamic>?)?.map((e) => e as String).toList()) ??
           ['linux'],
-      $checkedConvert(
-          json, 'dart', (v) => (v as List)?.map((e) => e as String)?.toList()),
+      $checkedConvert(json, 'dart',
+          (v) => (v as List<dynamic>?)?.map((e) => e as String).toList()),
       $checkedConvert(
           json,
           'stages',
-          (v) => (v as List)
-              ?.map((e) => e == null ? null : RawStage.fromJson(e as Map))
-              ?.toList()),
+          (v) => (v as List<dynamic>?)
+              ?.map((e) => RawStage.fromJson(e as Map))
+              .toList()),
       $checkedConvert(
           json, 'cache', (v) => v == null ? null : RawCache.fromJson(v as Map)),
     );
@@ -34,7 +34,7 @@ RawCache _$RawCacheFromJson(Map json) {
   return $checkedNew('RawCache', json, () {
     final val = RawCache(
       $checkedConvert(json, 'directories',
-          (v) => (v as List)?.map((e) => e as String)?.toList()),
+          (v) => (v as List<dynamic>).map((e) => e as String).toList()),
     );
     return val;
   });
